Ignore multiline comments and strings
It's relatively common for a multiline comment or string to not have
indentation on any but the first line, or to have unrelated indentation
from the surrounding code. If such a comment or string appeared in a
deeply indented location, it would interrupt the indent guides. Now,
lines that begin in the middle of a comment or string are ignored, so
guides are no longer interrupted in these cases.