Correct use of font lock
Previously, the use of font lock in this mode was a little hacky: it
would register a jit-lock function that would directly set the
font-lock-face text properties of the indents. It would also read said
properties from guides that were drawn earlier, in order to minimize the
amount of redrawing it would need to do. This behavior can be
problematic, and can potentially cause interference with other modes.

Now, this mode uses a custom text property to keep track of indentation
information, which is significantly less likely to cause problems. To
draw the guides, it now registers a font lock keyword that matches text
with this new property and applies the appropriate faces. This is the
correct way to use font lock and should not interfere with other modes.

Maybe it's a coincidence, but the mode seems to run faster now as well.

Also, all reference to the rear-nonsticky text property has been
removed, as it turns out it is no longer necessary.