Optimized performance with data memoization
Before, each indentation character had a highlight-indent-guides-prop text
property, containing 'even or 'odd, and a highlight-indent-guides-segment text
property, containing a list of indentation levels within that character. Now,
these have been consolidated into a single highlight-indent-guides-prop text
property. This property contains a list of all the useful indentation data for a
character, including the highlight level and segment data, as well as memoized
data for faster access during font-lock.

The memoized data includes the character's column width, and a 'starter' flag
representing whether the first column of the character is the first column of a
new indentation level. Both of these values used to be calculated during
font-lock, but it's more efficient to calculate them on modification and save
them for later. The memoized data also includes the most recent return value
from font-lock. This value, if it exists, is returned immediately from
font-lock, which is more efficient than recalculating it every time font-lock
runs. The value is deleted on modification, so that it can be recalculated on
the next font-lock cycle.

When jit-lock runs, it often refreshes lines that haven't been modified.
Re-applying the properties to these lines is unnecessary, and increasingly
inefficient. Now, before refreshing a line, the current indent guides will be
read from the line, and if they match the guides to be applied, the line is
skipped. This way, lines that won't change aren't recalculated.

It seems that pcase doesn't exist before emacs 24. Therefore, a dependency for
emacs 24 has been added. If this becomes a problem, it can be changed so pcase
isn't used. Also, font-lock-flush is used in place of font-lock-fontify-buffer,
as the latter is meant to be used interactively, not programmatically.