Fixed minor recalculation bug
The previous commit added a feature where the indent guides for a line
would only be reapplied if the guides to apply did not match the guides
that already exist on the line. This introduced a bug where, if you
insert non-whitespace characters before or within indentation, the line
would not be recalculated, and so the guides to the right of the new
character would not be cleared. Now, guides occurring after indentation
are always cleared, even when the indentation is not recalculated.