Fixed invisible indents bug
Inserting indentation manually (by pressing the spacebar) could
sometimes lead to the inserted spaces appearing invisible. This is due
to a combination of three behaviors:

1. When a character is typed, it inherits the properties from the
   preceding character, by default.
2. For performance reasons, this mode does not re-highlight lines that
   already appear to be highlighted correctly.
3. When a series of characters all have `eq' display properties, they
   are replaced with the display string as a group.

When a space is inserted after a space that is displayed as an indent
rule character, the properties are inherited, including the display
property. These display properties are not overwritten when the line is
still deemed "correct". This causes all such spaces to be combined, and
displayed as a single indent rule.

This is now fixed: the line reader now considers it invalid to have
either a starter character following a character of the same color, or a
non-starter character following a character of a different color
(before, it didn't consider the starter property at all). This causes
the offending lines to be redrawn.