Fixed issue with dynamic colors and daemon mode
Because of the way Emacs runs as a daemon, dynamic colors can't be
initialized on load. This is because at that point there is no frame
yet, so there are no theme faces to read colors from.

Now, dynamic colors are only initialized on load in standalone mode, and
they're also initialized each time a frame is created. This prevents the
daemon from printing an error on startup, and allows colors to set
properly when the first client connects.