Added responsive guides
Major rewrite multiple years in the making. Nearly doubled the line
count. Still shouldn't have taken this long, but took a long hiatus in
the middle due to life getting in the way.

Added the 'responsive' feature: highlights "current" guides with special
faces to show you which indent level you're on, like how Sublime does.
To make this work as efficiently as possible, refactored everything.

Added a way to override the default highlighter behavior by defining a
custom highlighter function.

I'm still not completely sold on the performance. We'll see if we can
make it faster somehow. For now this should be alright though.