Fixed empty update bug (#34)
When font-lock updates a range of zero characters in specific
situations (generally at the start of the buffer, or on an indented line
directly after a non-indented line), font-lock-fontify-region gets
called with a nil argument and throws an exception. Now, that argument
is checked beforehand and a reasonable alternative value is used
instead.