Responsive guide performance optimization
Before, to make responsive guides change colors appropriately, the
entire region containing said guides would be refontified. This is
inefficient, as it removes all fontification in the region and
recalculates it, even when that fontification has nothing to do with
indent guides. It also causes jit-lock to trigger, which reflows the
guides and then refontifies the region again. This is a lot of extra
work that just doesn't need to be done.

Now, a custom stripped-down function is called instead, which simply
draws guides on all guide characters in the region. No removal of
fontification, no calculating of unrelated fontification, no jit-lock
triggering. According to my ad-hoc perf tests, this runs about two or
three times as fast as the old implementation. The responsive delay is
still set to 0.1 seconds by default, but personally I wouldn't expect
most people to have any issues setting it to 0 anymore.

The responsive guides feature has proven itself over the past few days,
as far as I'm concerned. Removed the "unstable/unperformant" notice from
the readme.

The make-variable-buffer-local calls were also moved to the toplevel, to
prevent flycheck from complaining.