Fixed "splitting" bugs
A couple of bugs could occur when guides were "split" or "joined", by
inserting or pasting non-indented text between lines with a common
guide, or removing said text to cause separate-but-similar guides to
exist alongside each other on consecutive lines. The guides would not
split/join properly, which was apparent with responsive mode turned on.
Now, all split/join issues should be fixed.

I thought I spotted a single coloration glitch right after I wrote this
fix, but I could not get it to happen again, and after a good deal of
heavy use I was unable to spot another one. For now I'll assume I was
mistaken, but if I see it again I'll take steps to debug it.