Changed hosting for README images
Since adding the responsive gif five days ago, the Git LFS bandwidth for
my account exceeded 1.5GB and Git LFS has been disabled (I assume until
the end of the month, hopefully not longer). This is clearly due to the
responsive gif, which is stored in Git LFS. I expect this is because the
bandwidth limit counts every time someone views the README, which means
the README must have been getting a lot of views? I thought it might
have been the MELPA bot downloading the image every few hours, but
unless it wipes the repository every time, the image should be cached
and only downloaded once by Melpa. I can't think of any other source of
traffic that would cause this.

At any rate, I've uploaded the images to imgur and changed the README to
point to those. This should be a temporary solution while I try to come
up with a more reasonable permanent one, as I'd rather not rely on imgur
always being up, and would instead like to host these images on github
itself.