Compressed the linked animated gif
The gif has been compressed and re-linked. This is because Github
appears to be unable to display (cache?) the image, as it's too big. The
new compressed version is under a tenth of the original size, so while
the colors are a little off, it should now display properly. It should
also load more quickly than before.